requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0
django-environ==0.4.5
django-heroku==0.3.1
django-render-block==0.6
django-solo==1.1.3
django-storages==1.8
django-templated-email==2.3.0
dropbox==9.4.0
gunicorn==20.0.4
idna==2.8
Pillow==6.2.1
psycopg2==2.8.4
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==4.1.4